+++ cypress/integration/SearchForDuplicatesThroughAddRelationship/index.js
@@ -118,4 +118,3 @@ Then('all pagination is disabled', () => {
         .contains('Page 1')
         .should('exist');
 });
-