+++ src/core_modules/capture-core/components/DataEntryWidgetOutput/WarningsSection/SearchGroupDuplicate/WarningMessageCreator.types.js
@@ -1,0 +1,8 @@
+// @flow
+export type OwnProps = {|
+    onOpenReviewDialog: () => void,
+    dataEntryId: string,
+    selectedScopeId: string
+|}
+
+export type Props = {| ...OwnProps, ...CssClasses |}