+++ src/core_modules/capture-core/components/DataEntryWidgetOutput/WarningsSection/WaningsSection.container.js
@@ -1,8 +1,8 @@
 // @flow
 import { connect } from 'react-redux';
-import WarningsSection from '../../../../ViewEvent/RightColumn/WarningsSection/WarningsSection.component';
+import WarningsSection from '../../Pages/ViewEvent/RightColumn/WarningsSection/WarningsSection.component';
 import { makeGetSearchGroupWarning, makeGetWarningMessages } from './warningsSection.selectors';
-import { makeGetVisibleMessages } from '../../../../ViewEvent/RightColumn/ErrorsSection/messageSection.selectors';
+import { makeGetVisibleMessages } from '../../Pages/ViewEvent/RightColumn/ErrorsSection/messageSection.selectors';
 
 const makeStateToProps = () => {
     const getSearchGroupWarning = makeGetSearchGroupWarning();
@@ -31,6 +31,5 @@ const makeStateToProps = () => {
     return mapStateToProps;
 };
 
-// $FlowSuppress
 // $FlowFixMe[missing-annot] automated comment
 export default connect(makeStateToProps, () => ({}))(WarningsSection);