+++ src/core_modules/capture-core/components/Pages/New/NewPage.component.js
@@ -9,9 +9,10 @@ import { LockedSelector } from '../../LockedSelector';
 import type { ContainerProps, Props } from './NewPage.types';
 import { withErrorMessageHandler, withLoadingIndicator } from '../../../HOC';
 import { newPageStatuses } from './NewPage.constants';
-import { IncompleteSelectionsMessage } from '../../IncompleteSelectionsMessage';
 import { useScopeInfo } from '../../../hooks/useScopeInfo';
 import { RegistrationDataEntry } from './RegistrationDataEntry';
+import { NoWriteAccessMessage } from '../../NoWriteAccessMessage';
+import { IncompleteSelectionsMessage } from '../../IncompleteSelectionsMessage';
 
 const getStyles = () => ({
     container: {
@@ -29,6 +30,7 @@ const NewPagePlain = ({
     classes,
     currentScopeId,
     newPageStatus,
+    writeAccess,
     programCategorySelectionIncomplete,
     missingCategoriesInProgramSelection,
     orgUnitSelectionIncomplete,
@@ -61,45 +63,56 @@ const NewPagePlain = ({
         <LockedSelector pageToPush="new" />
         <div data-test="dhis2-capture-registration-page-content" className={classes.container} >
             {
-                newPageStatus === newPageStatuses.DEFAULT &&
-                <RegistrationDataEntry
-                    dataEntryId={NEW_TEI_DATA_ENTRY_ID}
-                    selectedScopeId={selectedScopeId}
-                    setScopeId={setScopeId}
-                />
-            }
-
-            {
-                newPageStatus === newPageStatuses.WITHOUT_ORG_UNIT_SELECTED &&
-                <>
-                    <IncompleteSelectionsMessage>
-                        {i18n.t('Choose a registering unit to start reporting')}
-                    </IncompleteSelectionsMessage>
-                    <Button
-                        dataTest="dhis2-capture-new-page-cancel-button"
-                        onClick={handleMainPageNavigation}
-                    >
-                        {i18n.t('Cancel')}
-                    </Button>
-                </>
-            }
-
-            {
-                newPageStatus === newPageStatuses.WITHOUT_PROGRAM_CATEGORY_SELECTED &&
-                (() => {
-                    const missingCategories = missingCategoriesInProgramSelection.reduce((acc, { name }, index) => {
-                        if ((index + 1 === missingCategoriesInProgramSelection.length)) {
-                            return `${acc} ${name} ${missingCategoriesInProgramSelection.length > 1 ? 'categories' : 'category'}`;
+                !writeAccess ?
+                    <NoWriteAccessMessage
+                        title={i18n.t('New')}
+                        message={i18n.t("You don't have access to create an event in the current selections")}
+                    />
+                    :
+                    <>
+                        {
+                            newPageStatus === newPageStatuses.DEFAULT &&
+                            <RegistrationDataEntry
+                                dataEntryId={NEW_TEI_DATA_ENTRY_ID}
+                                selectedScopeId={selectedScopeId}
+                                setScopeId={setScopeId}
+                            />
                         }
-                        return `${acc} ${name},`;
-                    }, '');
 
-                    return (
-                        <IncompleteSelectionsMessage>
-                            {i18n.t('Choose the {{missingCategories}} to start reporting', { missingCategories })}
-                        </IncompleteSelectionsMessage>
-                    );
-                })()
+                        {
+                            newPageStatus === newPageStatuses.WITHOUT_ORG_UNIT_SELECTED &&
+                            <>
+                                <IncompleteSelectionsMessage>
+                                    {i18n.t('Choose a registering unit to start reporting')}
+                                </IncompleteSelectionsMessage>
+                                <Button
+                                    dataTest="dhis2-capture-new-page-cancel-button"
+                                    onClick={handleMainPageNavigation}
+                                >
+                                    {i18n.t('Cancel')}
+                                </Button>
+                            </>
+                        }
+
+                        {
+                            newPageStatus === newPageStatuses.WITHOUT_PROGRAM_CATEGORY_SELECTED &&
+                            (() => {
+                                const missingCategories = missingCategoriesInProgramSelection.reduce((acc, { name }, index) => {
+                                    if ((index + 1 === missingCategoriesInProgramSelection.length)) {
+                                        return `${acc} ${name} ${missingCategoriesInProgramSelection.length > 1 ? 'categories' : 'category'}`;
+                                    }
+                                    return `${acc} ${name},`;
+                                }, '');
+
+                                return (
+                                    <IncompleteSelectionsMessage>
+                                        {i18n.t('Choose the {{missingCategories}} to start reporting', { missingCategories })}
+                                    </IncompleteSelectionsMessage>
+                                );
+                            })()
+                        }
+
+                    </>
             }
         </div>
     </>);