+++ src/core_modules/capture-core/components/Pages/New/RegistrationDataEntry/RegistrationDataEntry.container.js
@@ -7,6 +7,7 @@ import {
     startSavingNewTrackedEntityInstance,
     startSavingNewTrackedEntityInstanceWithEnrollment,
 } from './RegistrationDataEntry.actions';
+import { useDuplicates } from '../../../PossibleDuplicatesDialog/useDuplicates';
 
 export const RegistrationDataEntry: ComponentType<OwnProps>
   = ({ selectedScopeId, dataEntryId, setScopeId }) => {
@@ -20,6 +21,7 @@ export const RegistrationDataEntry: ComponentType<OwnProps>
           () => { dispatch(startSavingNewTrackedEntityInstanceWithEnrollment()); },
           [dispatch]);
 
+      const { onReviewDuplicates } = useDuplicates(dataEntryId, selectedScopeId);
 
       const dataEntryIsReady = useSelector(({ dataEntries }) => (!!dataEntries[dataEntryId]));
 
@@ -29,6 +31,7 @@ export const RegistrationDataEntry: ComponentType<OwnProps>
               selectedScopeId={selectedScopeId}
               setScopeId={setScopeId}
               dataEntryIsReady={dataEntryIsReady}
+              onReviewDuplicates={onReviewDuplicates}
               onSaveWithoutEnrollment={dispatchOnSaveWithoutEnrollment}
               onSaveWithEnrollment={dispatchOnSaveWithEnrollment}
           />);