+++ src/core_modules/capture-core/components/PossibleDuplicatesDialog/ReviewDialogContents/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { ReviewDialogContents } from './ReviewDialogContents.container';